+++ window-utils-test/window-utils-test.c
@@ -51,7 +51,7 @@ TEST t_windows_search(void *NAME){
 TEST t_authorized_tests(void){
   authorized_test_t *authorized_test_results = execute_authorization_tests();
 
-  for (int i = 0; i < (sizeof(*authorized_test_results) / sizeof((authorized_test_results)[0])); i++) {
+  for (size_t i = 0; i < (sizeof(*authorized_test_results) / sizeof((authorized_test_results)[0])); i++) {
     ASSERT_EQm(authorized_test_results[i].name, authorized_test_results[i].authorized, true);
   }
 }
